adding refactored e2es
