--- cypress/test-data/worldServicesFpData.js
@@ -1,20 +1,0 @@
-export default {
-  igbo: {
-    data: {
-      h2: 'Isi akụkọ',
-      h2Selector: '#Top-stories',
-    },
-  },
-  pidgin: {
-    data: {
-      h2: 'Top tori',
-      h2Selector: '#Top-stories',
-    },
-  },
-  yoruba: {
-    data: {
-      h2: 'Ìròyìn tó ṣe kókó',
-      h2Selector: '#Top-stories',
-    },
-  },
-};
